#12[ui] chip 디자인 styles.xml에 지정
