+++ app/src/main/res/layout/fragment_search.xml
@@ -120,11 +120,11 @@
         android:layout_width="match_parent"
         android:layout_height="0dp"
         android:layout_marginTop="20dp"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/tv_searchfilter_addr"
-        tools:listitem="@layout/constraint_layout_search_rv"/>
+        tools:listitem="@layout/constraint_layout_search_rv" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
