+++ app/src/main/res/layout/search_bottomsheet_dialog_children.xml
@@ -38,9 +38,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/filter_children_possible"
-            android:textSize="16sp"
             app:chipEndPadding="30dp"
-            app:chipMinHeight="50dp"
             app:chipStartPadding="30dp"
             app:layout_constraintEnd_toStartOf="@+id/cp_bottom_children_impossible"
             app:layout_constraintHorizontal_bias="0.5"
@@ -54,9 +52,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/filter_children_impossible"
-            android:textSize="16sp"
             app:chipEndPadding="30dp"
-            app:chipMinHeight="50dp"
             app:chipStartPadding="30dp"
             app:layout_constraintBottom_toBottomOf="@id/cp_bottom_children_possible"
             app:layout_constraintEnd_toEndOf="parent"
