+++ app/src/main/res/layout/search_bottomsheet_dialog_genre.xml
@@ -24,10 +24,9 @@
         android:id="@+id/cp_bottom_genre_theater"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="30dp"
         android:text="@string/filter_genre_theater"
-        android:textSize="16sp"
         app:layout_constraintEnd_toStartOf="@+id/cp_bottom_futility"
         app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintHorizontal_chainStyle="packed"
@@ -38,10 +37,9 @@
         android:id="@+id/cp_bottom_futility"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_futility"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_genre_theater"
         app:layout_constraintEnd_toStartOf="@+id/cp_bottom_publicfutility"
         app:layout_constraintHorizontal_bias="0.5"
@@ -52,10 +50,9 @@
         android:id="@+id/cp_bottom_publicfutility"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_publicfutility"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_futility"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintHorizontal_bias="0.5"
@@ -66,10 +63,9 @@
         android:id="@+id/cp_bottom_classic"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="5dp"
         android:text="@string/filter_genre_classic"
-        android:textSize="16sp"
         app:layout_constraintStart_toStartOf="@id/cp_bottom_genre_theater"
         app:layout_constraintTop_toBottomOf="@+id/cp_bottom_genre_theater" />
 
@@ -77,10 +73,9 @@
         android:id="@+id/cp_bottom_korean_classic"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_korean_classic"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_classic"
         app:layout_constraintStart_toEndOf="@+id/cp_bottom_classic"
         app:layout_constraintTop_toTopOf="@+id/cp_bottom_classic" />
@@ -89,10 +84,9 @@
         android:id="@+id/cp_bottom_popular_music"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_popular_music"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_korean_classic"
         app:layout_constraintStart_toEndOf="@+id/cp_bottom_korean_classic"
         app:layout_constraintTop_toTopOf="@+id/cp_bottom_korean_classic" />
@@ -101,10 +95,9 @@
         android:id="@+id/cp_bottom_musical"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginTop="5dp"
         android:text="@string/filter_genre_musical"
-        android:textSize="16sp"
         app:layout_constraintStart_toStartOf="@id/cp_bottom_classic"
         app:layout_constraintTop_toBottomOf="@+id/cp_bottom_classic" />
 
@@ -112,10 +105,9 @@
         android:id="@+id/cp_bottom_circus"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_circus"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_musical"
         app:layout_constraintStart_toEndOf="@+id/cp_bottom_musical"
         app:layout_constraintTop_toTopOf="@+id/cp_bottom_musical" />
@@ -124,10 +116,9 @@
         android:id="@+id/cp_bottom_mix"
         style="@style/CustomChipChoice."
         android:layout_width="100dp"
-        android:layout_height="40dp"
+        android:layout_height="wrap_content"
         android:layout_marginStart="5dp"
         android:text="@string/filter_genre_mix"
-        android:textSize="16sp"
         app:layout_constraintBottom_toBottomOf="@+id/cp_bottom_circus"
         app:layout_constraintStart_toEndOf="@+id/cp_bottom_circus"
         app:layout_constraintTop_toTopOf="@+id/cp_bottom_circus" />
