+++ app/src/main/res/values/styles.xml
@@ -3,6 +3,9 @@
     <style name="CustomChipChoice." parent="@style/Widget.MaterialComponents.Chip.Choice">
         <item name="chipBackgroundColor" >@color/search_chip_bg_state_list</item>
         <item name="android:textColor" >@color/search_chip_text_state_list</item>
+        <item name="android:textSize">14sp</item>
+        <item name="chipCornerRadius">5dp</item>
+        <item name="chipMinHeight">45dp</item>
     </style>
 
 </resources>
